Adjust package structure and relocate OrderTicket, Order, Client, and Product classes
