+++ src/main/java/com/tcc/serveme/api/dto/OrderRequest.java
@@ -1,7 +1,7 @@
 package com.tcc.serveme.api.dto;
 
-import com.tcc.serveme.api.model.OrderItem;
+import com.tcc.serveme.api.model.OrderTicket;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record OrderRequest(int tableNumber, String waiter, List<OrderItem> items, String notes, LocalDateTime createdAt) {}
+public record OrderRequest(int tableNumber, String waiter, List<OrderTicket> items, String notes, LocalDateTime createdAt) {}
