--- src/main/java/com/tcc/serveme/api/model/OrderItem.java
@@ -1,31 +1,0 @@
-package com.tcc.serveme.api.model;
-
-public class OrderItem {
-    private String productName;
-    private int quantity;
-    private double unitPrice;
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public void setProductName(String productName) {
-        this.productName = productName;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public double getUnitPrice() {
-        return unitPrice;
-    }
-
-    public void setUnitPrice(double unitPrice) {
-        this.unitPrice = unitPrice;
-    }
-}
