+++ src/main/java/com/tcc/serveme/api/model/OrderTicket.java
@@ -1,4 +1,4 @@
-package com.tcc.serveme.model.entities;
+package com.tcc.serveme.api.model;
 
 import java.util.ArrayList;
 import java.util.List;
