--- src/main/java/com/tcc/serveme/model/entities/Order.java
@@ -1,33 +1,0 @@
-package com.tcc.serveme.model.entities;
-
-public class Order {
-
-    private Product product;
-    private Integer quantity;
-
-    public Order(Product product, Integer quantity) {
-        this.product = product;
-        this.quantity = quantity;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public void setProduct(Product product) {
-        this.product = product;
-    }
-
-    public Integer getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(Integer quantity) {
-        this.quantity = quantity;
-    }
-
-    public Double subTotal(){
-        return quantity * product.getPrice();
-    }
-
-}
