--- src/main/java/com/tcc/serveme/model/enums/Status.java
@@ -1,8 +1,0 @@
-package com.tcc.serveme.model.enums;
-
-public enum Status {
-
-    OPEN,
-    CLOSED;
-
-}
