--- src/main/java/com/tcc/serveme/model/exceptions/DomainException.java
@@ -1,14 +1,0 @@
-package com.tcc.serveme.model.exceptions;
-
-public class DomainException extends Exception{ 
-/* 
-    É possivel extender da classe RunTimeException caso não tenha 
-    necessidade de tratar a excessão para rodar o código e caso ela venha a ocorrer o código irá quebrar.
-*/
-
-    private static final long serialVersionUID = 1L;
-
-    public DomainException(String msg){
-        super(msg);
-    }
-}
